Fix SLO objective float32 precision loss (issue #2396)
PR #2323 migrated the SLO resource to the shared kbapi client, which
changed SLOsObjective.Target and similar fields from float64 to float32
because the Kibana OpenAPI spec uses `type: number` without `format`.
This caused `Provider produced inconsistent result after apply` errors:
float64(float32(0.999)) = 0.9990000128746033, not 0.999.

Fix: add a `fixSloFloatFormats` transformer in transform_schema.go that
sets `format: double` on all SLO floating-point fields before codegen.
oapi-codegen then emits float64 instead of float32, preserving full
IEEE-754 double precision through the read/write cycle.

Fields fixed: SLOs_objective target/timesliceTarget, histogram good/total
from/to, APM latency threshold, timeslice metric threshold and percentile.

Remove the now-unnecessary float32ToFloat64, float64PtrToFloat32Ptr, and
float32PtrToFloat64Ptr conversion helpers, and add an acceptance test
(TestAccResourceSloFloatPrecision) that verifies 0.999 round-trips
without corruption.

+++ internal/kibana/slo/models_apm_latency_indicator.go
@@ -50,7 +50,7 @@ func (m tfModel) apmLatencyIndicatorToAPI() (bool, kbapi.SLOsSloWithSummaryRespo
 			Filter          *string `json:"filter,omitempty"`
 			Index           string  `json:"index"`
 			Service         string  `json:"service"`
-			Threshold       float32 `json:"threshold"`
+			Threshold       float64 `json:"threshold"`
 			TransactionName string  `json:"transactionName"`
 			TransactionType string  `json:"transactionType"`
 		}{
@@ -60,7 +60,7 @@ func (m tfModel) apmLatencyIndicatorToAPI() (bool, kbapi.SLOsSloWithSummaryRespo
 			TransactionName: ind.TransactionName.ValueString(),
 			Filter:          stringPtr(ind.Filter),
 			Index:           ind.Index.ValueString(),
-			Threshold:       float32(ind.Threshold.ValueInt64()),
+			Threshold:       float64(ind.Threshold.ValueInt64()),
 		},
 	}
